sorrentum_sandbox/spring2024/sorrentum_sandbox/spring2023/ml_projects/Issue25_Team6_Implement_sandbox_for_Bitquery_and_Uniswap/scripts/requirements.txt
install
redis
requests
pandas
numpy
matplotlib
seaborn
notebook
jupyter
apache-airflow[postgres]
psycopg2-binary
python-dotenv
